Revert "Fix off_t to int truncation in bio repl transfer size reporting (#3811)"
This reverts commit b4ea71ee19c1e8af3177c7df3bce229e35359aba.

+++ src/replication.c
@@ -5221,8 +5221,8 @@ void handleBioThreadFinishedRDBDownload(void) {
     debugServerAssert(bio_save_state == REPL_BIO_DISK_SAVE_STATE_FINISHED);
 
     /* Bio termination detected - we can get rid of the state vars */
-    off_t bio_repl_transfer_size = server.bio_repl_transfer_size;
-    off_t bio_repl_transfer_read = server.bio_repl_transfer_read;
+    int bio_repl_transfer_size = server.bio_repl_transfer_size;
+    int bio_repl_transfer_read = server.bio_repl_transfer_read;
     resetBioRDBSaveState();
 
     serverLog(LL_NOTICE, "Replica main thread detected RDB download completion in Bio thread");
